requirements.txt
flask>=3.0.0
importlib-metadata
nest-asyncio
pandas>=1.2.3
plotly>=5.0.0
requests
retrying
setuptools
typing-extensions>=4.1.1

# Core Dash 3.x (components now included in dash package)
dash>=3.0.0
dash-iconify>=0.1.2
dash-mantine-components>=2.4.0
dash-ag-grid
dash-dock
dash-resizable-panels

# Documentation & Utilities
requests>=2.27.1
python-frontmatter>=1.0.0
markdown2dash
pydantic>=2.3.0

# AI/LLM Integration & SEO
dash-improve-my-llms>=0.3.0
anthropic>=0.39.0
python-dotenv>=1.0.0

# Production Server
gunicorn>=21.2.0
